Make renderer file
